requirements.txt
discord.py
wavelink